Authorization (deleting pet)

+++ src/actions/actions.ts
@@ -41,6 +41,7 @@ export async function logOut() {
 export async function addPet(pet: unknown) {
   await sleep(1000);
 
+  // authentication check
   const session = await auth();
   if (!session?.user) {
     redirect('/login');
@@ -103,6 +104,13 @@ export async function editPet(petId: unknown, newPetData: unknown) {
 export async function deletePet(petId: unknown) {
   await sleep(1000);
 
+  // authentication check
+  const session = await auth();
+  if (!session?.user) {
+    redirect('/login');
+  }
+
+  // validation
   const validatedPetId = petIdSchema.safeParse(petId);
 
   if (!validatedPetId.success) {
@@ -110,6 +118,30 @@ export async function deletePet(petId: unknown) {
       messsage: 'Invalid pet data.',
     };
   }
+
+  // authorization check (user owns pet)
+  const pet = await prisma.pet.findUnique({
+    where: {
+      id: validatedPetId.data,
+    },
+    select: {
+      userId: true,
+    },
+  });
+
+  if (!pet) {
+    return {
+      message: 'Pet not found.',
+    };
+  }
+
+  if (pet.userId !== session.user.id) {
+    return {
+      message: 'Not authorized.',
+    };
+  }
+
+  // database mutation
   try {
     await prisma.pet.delete({
       where: {
